support whatwg streams in decodeAsync() and decodeArrayStream()
