+++ src/decodeAsync.ts
@@ -1,13 +1,18 @@
 import { Decoder } from "./Decoder";
 import { defaultDecodeOptions, DecodeOptions } from "./decode";
+import { isReadableStream, asyncIterableFromStream } from "./utils/stream";
 
 export type DecodeAsyncOptions = DecodeOptions;
 export const defaultDecodeAsyncOptions = defaultDecodeOptions;
 
+type StreamLike<T> = AsyncIterable<T> | ReadableStream<T>;
+
 export async function decodeAsync(
-  stream: AsyncIterable<Uint8Array | ArrayLike<number>>,
+  streamLike: StreamLike<Uint8Array | ArrayLike<number>>,
   options: DecodeAsyncOptions = defaultDecodeOptions,
 ): Promise<unknown> {
+  const stream = isReadableStream(streamLike) ? asyncIterableFromStream(streamLike) : streamLike;
+
   const decoder = new Decoder(
     options.extensionCodec,
     options.maxStrLength,
@@ -20,9 +25,11 @@ export async function decodeAsync(
 }
 
 export async function* decodeArrayStream(
-  stream: AsyncIterable<Uint8Array | ArrayLike<number>>,
+  streamLike: StreamLike<Uint8Array | ArrayLike<number>>,
   options: DecodeAsyncOptions = defaultDecodeOptions,
 ) {
+  const stream = isReadableStream(streamLike) ? asyncIterableFromStream(streamLike) : streamLike;
+
   const decoder = new Decoder(
     options.extensionCodec,
     options.maxStrLength,
